+++ src/main/java/com/mapshot/api/admin/service/AdminService.java
@@ -2,8 +2,8 @@ package com.mapshot.api.admin.service;
 
 import com.mapshot.api.admin.model.AdminRequest;
 import com.mapshot.api.admin.repository.AdminRepository;
-import com.mapshot.api.common.exception.ApiException;
-import com.mapshot.api.common.exception.status.ErrorCode;
+import com.mapshot.api.infra.web.exception.ApiException;
+import com.mapshot.api.infra.web.exception.status.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
