--- src/main/java/com/mapshot/api/auth/annotation/PreAuth.java
@@ -1,14 +1,0 @@
-package com.mapshot.api.auth.annotation;
-
-import com.mapshot.api.auth.enums.Accessible;
-
-import java.lang.annotation.*;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface PreAuth {
-
-    Accessible[] value();
-
-}
