--- src/main/java/com/mapshot/api/auth/config/WebAuthConfig.java
@@ -1,43 +1,0 @@
-package com.mapshot.api.auth.config;
-
-import com.mapshot.api.auth.interceptor.AuthInterceptor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class WebAuthConfig implements WebMvcConfigurer {
-
-    private final AuthInterceptor authInterceptor;
-
-    @Value("${jwt.admin.header}")
-    public String ADMIN_HEADER_NAME;
-
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-
-        registry.addInterceptor(authInterceptor)
-                .order(1)
-                .excludePathPatterns("/css/**")
-                .excludePathPatterns("/js/**")
-                .excludePathPatterns("/images/**")
-                .excludePathPatterns("/error/**")
-                .excludePathPatterns("/download/**")
-                .excludePathPatterns("/favicon.ico")
-                .addPathPatterns("/**");
-    }
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/admin/**")
-                .exposedHeaders(ADMIN_HEADER_NAME)
-                .allowCredentials(true);
-
-    }
-
-}
