--- src/main/java/com/mapshot/api/common/exception/status/StatusCode.java
@@ -1,9 +1,0 @@
-package com.mapshot.api.common.exception.status;
-
-import org.springframework.http.HttpStatus;
-
-public interface StatusCode {
-    HttpStatus getHttpStatus();
-
-    String getMessage();
-}
