--- src/main/java/com/mapshot/api/common/filter/RequestWrapper.java
@@ -1,55 +1,0 @@
-package com.mapshot.api.common.filter;
-
-
-import jakarta.servlet.ReadListener;
-import jakarta.servlet.ServletInputStream;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletRequestWrapper;
-import org.springframework.util.StreamUtils;
-
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-
-public class RequestWrapper extends HttpServletRequestWrapper {
-
-    private final byte[] cachedInputStream;
-
-    public RequestWrapper(HttpServletRequest request) throws IOException {
-        super(request);
-        InputStream requestInputStream = request.getInputStream();
-        this.cachedInputStream = StreamUtils.copyToByteArray(requestInputStream);
-    }
-
-    @Override
-    public ServletInputStream getInputStream() {
-        return new ServletInputStream() {
-            private final InputStream cachedBodyInputStream = new ByteArrayInputStream(cachedInputStream);
-
-            @Override
-            public boolean isFinished() {
-                try {
-                    return cachedBodyInputStream.available() == 0;
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                return false;
-            }
-
-            @Override
-            public boolean isReady() {
-                return true;
-            }
-
-            @Override
-            public void setReadListener(ReadListener readListener) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public int read() throws IOException {
-                return cachedBodyInputStream.read();
-            }
-        };
-    }
-}
