--- src/main/java/com/mapshot/api/common/filter/ResponseWrapper.java
@@ -1,11 +1,0 @@
-package com.mapshot.api.common.filter;
-
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.web.util.ContentCachingResponseWrapper;
-
-
-public class ResponseWrapper extends ContentCachingResponseWrapper {
-    public ResponseWrapper(HttpServletResponse response) {
-        super(response);
-    }
-}
