--- src/main/java/com/mapshot/api/common/slack/model/ErrorMessage.java
@@ -1,15 +1,0 @@
-package com.mapshot.api.common.slack.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ErrorMessage {
-    private String title;
-    private String message;
-}
