+++ src/main/java/com/mapshot/api/image/controller/MapTemplateController.java
@@ -1,8 +1,8 @@
 package com.mapshot.api.image.controller;
 
-import com.mapshot.api.auth.annotation.PreAuth;
-import com.mapshot.api.auth.enums.Accessible;
-import com.mapshot.api.image.model.ImageRequest;
+import com.mapshot.api.infra.client.lambda.LambdaRequest;
+import com.mapshot.api.infra.web.auth.annotation.PreAuth;
+import com.mapshot.api.infra.web.auth.enums.Accessible;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,7 +15,7 @@ public class MapTemplateController {
 
     @PreAuth(Accessible.FRIENDLY_SERVER)
     @GetMapping("/kakao")
-    public String getKakaoMap(@ModelAttribute ImageRequest mapRequest, Model model) {
+    public String getKakaoMap(@ModelAttribute LambdaRequest mapRequest, Model model) {
         model.addAttribute("mapRequest", mapRequest);
 
         return "map/kakao";
@@ -23,7 +23,7 @@ public class MapTemplateController {
 
     @PreAuth(Accessible.FRIENDLY_SERVER)
     @GetMapping("/google")
-    public String getGoogleMap(@ModelAttribute ImageRequest mapRequest, Model model) {
+    public String getGoogleMap(@ModelAttribute LambdaRequest mapRequest, Model model) {
         model.addAttribute("mapRequest", mapRequest);
 
         return "map/google";
