--- src/main/java/com/mapshot/api/image/model/ImageResponse.java
@@ -1,17 +1,0 @@
-package com.mapshot.api.image.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ImageResponse {
-
-    private int x;
-    private int y;
-    private String uuid;
-}
