+++ src/main/java/com/mapshot/api/infra/client/slack/model/SlackMessage.java
@@ -1,0 +1,15 @@
+package com.mapshot.api.infra.client.slack.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class SlackMessage {
+    private String title;
+    private String message;
+}
