+++ src/main/java/com/mapshot/api/infra/web/exception/ApiException.java
@@ -1,0 +1,24 @@
+package com.mapshot.api.infra.web.exception;
+
+import com.mapshot.api.infra.web.exception.status.StatusCode;
+import lombok.Getter;
+
+@Getter
+public class ApiException extends RuntimeException {
+
+    private StatusCode code;
+
+    public ApiException(String msg) {
+        super(msg);
+    }
+
+    public ApiException(StatusCode code) {
+        super(code.getMessage());
+        this.code = code;
+    }
+
+    public ApiException(String msg, Throwable e) {
+        super(msg, e);
+    }
+
+}
