+++ src/main/java/com/mapshot/api/infra/web/exception/status/StatusCode.java
@@ -1,0 +1,9 @@
+package com.mapshot.api.infra.web.exception.status;
+
+import org.springframework.http.HttpStatus;
+
+public interface StatusCode {
+    HttpStatus getHttpStatus();
+
+    String getMessage();
+}
