+++ src/main/java/com/mapshot/api/infra/web/exception/status/SuccessCode.java
@@ -1,0 +1,19 @@
+package com.mapshot.api.infra.web.exception.status;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public enum SuccessCode implements StatusCode {
+
+    OK(HttpStatus.OK, "ok");
+
+    private final HttpStatus httpStatus;
+    private final String message;
+
+    private SuccessCode(HttpStatus httpStatus, String message) {
+        this.httpStatus = httpStatus;
+        this.message = message;
+    }
+
+}
