+++ src/main/java/com/mapshot/api/infra/web/logging/LoggingFilter.java
@@ -1,0 +1,86 @@
+package com.mapshot.api.infra.web.logging;
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.slf4j.MDC;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Component;
+import org.springframework.util.StreamUtils;
+import org.springframework.web.filter.OncePerRequestFilter;
+import org.springframework.web.util.ContentCachingResponseWrapper;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Arrays;
+import java.util.List;
+import java.util.UUID;
+
+@Component
+@Slf4j
+public class LoggingFilter extends OncePerRequestFilter {
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        MDC.put("traceId", UUID.randomUUID().toString());
+        if (isAsyncDispatch(request)) {
+            filterChain.doFilter(request, response);
+        } else {
+            doFilterWrapped(new RequestWrapper(request), new ResponseWrapper(response), filterChain);
+        }
+        MDC.clear();
+    }
+
+    protected void doFilterWrapped(RequestWrapper request, ContentCachingResponseWrapper response, FilterChain filterChain) throws ServletException, IOException {
+        try {
+            logRequest(request);
+            filterChain.doFilter(request, response);
+        } finally {
+            logResponse(response);
+            response.copyBodyToResponse();
+        }
+    }
+
+    private static void logRequest(RequestWrapper request) throws IOException {
+        String queryString = request.getQueryString();
+        log.info("Request : {} [{}]",
+                request.getMethod(),
+                queryString == null ? request.getRequestURI() : request.getRequestURI() + queryString);
+
+//        logPayload("Request", request.getContentType(), request.getInputStream());
+    }
+
+    private static void logResponse(ContentCachingResponseWrapper response) {
+        log.info("Response : {}", response.getStatus());
+    }
+
+    private static void logPayload(String prefix, String contentType, InputStream inputStream) throws IOException {
+        boolean visible = isVisible(MediaType.valueOf(contentType == null ? "application/json" : contentType));
+        if (visible) {
+            byte[] content = StreamUtils.copyToByteArray(inputStream);
+            if (content.length > 0) {
+                String contentString = new String(content);
+                log.info("{} Payload: {}", prefix, contentString);
+            }
+        } else {
+            log.info("{} Payload: Binary Content", prefix);
+        }
+    }
+
+    private static boolean isVisible(MediaType mediaType) {
+        final List<MediaType> VISIBLE_TYPES = Arrays.asList(
+                MediaType.valueOf("text/*"),
+                MediaType.APPLICATION_FORM_URLENCODED,
+                MediaType.APPLICATION_JSON,
+                MediaType.APPLICATION_XML,
+                MediaType.valueOf("application/*+json"),
+                MediaType.valueOf("application/*+xml"),
+                MediaType.MULTIPART_FORM_DATA
+        );
+
+        return VISIBLE_TYPES.stream()
+                .anyMatch(visibleType -> visibleType.includes(mediaType));
+    }
+}
