+++ src/main/java/com/mapshot/api/infra/web/logging/ResponseWrapper.java
@@ -1,0 +1,11 @@
+package com.mapshot.api.infra.web.logging;
+
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.web.util.ContentCachingResponseWrapper;
+
+
+public class ResponseWrapper extends ContentCachingResponseWrapper {
+    public ResponseWrapper(HttpServletResponse response) {
+        super(response);
+    }
+}
