+++ src/test/java/com/mapshot/api/image/controller/StorageControllerTest.java
@@ -3,9 +3,9 @@ package com.mapshot.api.image.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mapshot.api.SlackMockExtension;
-import com.mapshot.api.auth.validation.Validation;
 import com.mapshot.api.image.model.StorageRequest;
 import com.mapshot.api.image.service.StorageService;
+import com.mapshot.api.infra.web.auth.Validation;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
